updating file structure
